We used to constantly update a segment's size, though we didn't force the update to disk. However, reading up on msync, it seems like a lot (most?) implementations will take the opportunity to sync all changed buffers. This is unfortunate since the size was in the head, while data (which we do care to sync) is at the tail.
With this new approach, the size is only set when we're finished with a segment.
we're already sync'ing the head of the file to persist the next segment id,
so we get this sync for free.

This change continues to guarantee that archived segments have a proper size
(and hopefully the locking is correct and this guarantee holds for a newly
archived segment).

The only downside is that is that for an unsized segment (which is the current
segment) we have to scan the entire segment to find the correct position. This,
thankfully, only happens when restarting a topic.